spe-parser: add prototype tool to parse SPE data
spe-parser is a tool used to parse SPE data from `perf` raw data that
contains SPE records. It will save parsed SPE records into files with
parquet format.

This tool is an early prototype. It's scheduled for a re-write, possibly
in python, and the production version will be very different from this
early version.

Tested-by: Rui Chang <rui.chang@arm.com>
Co-Authored-by: Rui Chang <rui.chang@arm.com>
Change-Id: Ia33f0044ccf3613201b2832aa418d06ff382096f